// src/utils/enums.ts
export enum ICON_TOOL_ACTIVE {
  LAYER = 'LAYER',
  BOOKMARK = 'BOOKMARK',
  PRINTING = 'PRINTING',
  ZOOM_IN = 'ZOOM_IN',
  ZOOM_OUT = 'ZOOM_OUT',
  SEARCH = 'SEARCH',
  FULL_SCREEN = 'FULL_SCREEN',
  POLYGON = 'POLYGON',
  CARING_AREA = 'CARING_AREA',
  // MEASURING_VERTICAL = 'MEASURING_VERTICAL',
  MEASURING_HORIZONTAL = 'MEASURING_HORIZONTAL',
  RULER = 'RULER',
  UPLOAD = 'UPLOAD',
  SCENE3D = 'SCENE3D',
  SCENE2D = 'SCENE2D',
  CHANGE_DETECTION_INFO = 'CHANGE_DETECTION_INFO',
}

export enum FILE_TYPE {
  SHP = '.shp',
  SHX = '.shx',
  DBF = '.dbf',
  PRJ = '.prj',
  XML = '.xml',
}

export enum ID_MAP {
  COMPARE_MAP = 'COMPARE_MAP',
}

export enum LOCAL_STORAGE_KEY {
  ACCESS_TOKEN = 'ACCESS_TOKEN',
  USER_INFO = 'USER_INFO',
  LOCALE = 'LOCALE',
}

export enum MENU_KEY {
  MENU = 'MENU',
  APP = 'APP',
  USER = 'USER',
  ROLE = 'ROLE',
  POSITION = 'POSITION',
  RIGHT = 'RIGHT',
  UNIT = 'UNIT',
  LOG = 'LOG',
  DATA_TYPE = 'DATA_TYPE',
  FIELD = 'FIELD',
  ADMINISTRATIVE_UNITS = 'ADMINISTRATIVE_UNITS',
}

export enum ROLE {
  VIEWER = 'VIEWER',
  USER = 'USER',
  ADMIN = 'ADMIN',
}

export enum SCOPE {
  ADMIN_EDIT = 'ADMIN.EDIT',
}

export enum GROUP_LAYER_TYPE {
  MAC_DINH = 'MAC_DINH', // LAYER_MAP
  VE_TINH = 'VE_TINH', // LAYER_MAP

  HANH_CHINH = 'HANH_CHINH', //
  CHUYEN_NGANH = 'CHUYEN_NGANH', //
  LOP_PHU = 'LOP_PHU', //,
  DOI_TUONG = 'DOI_TUONG', //,
}

export enum DRAW_TYPE {
  POLYGON = 'POLYGON',
  MEASURE_POLYGON = 'MEASURE_POLYGON',
}

export enum RIGHT_BOTTOM {
  COMPARE = 'COMPARE',
  DISPLAY = 'DISPLAY',
  DETECT = 'DETECT',
}

export enum ACTIVE_FUNCTION {
  SEARCH_IMAGES = 'SEARCH_IMAGES',
  DETECT_CHANGE = 'DETECT_CHANGE',
  DETECT_OBJECT = 'DETECT_OBJECT',
  MONITOR_CURRENT_STATUS = 'MONITOR_CURRENT_STATUS',
  DEFAULT = 'DEFAULT',
}